requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0
cloud-sql-python-connector>=1.2.0
google-auth>=2.22.0
pg8000>=1.29.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# API and CORS
starlette>=0.27.0
python-multipart>=0.0.6
httpx>=0.25.0

# Date/Time handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0

# Development
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=0.991